moves enquiying logic to orchestrator
